screenresolution fixed to 16:9

+++ src/main/java/com/softpro/dnaig/softpro/MainApplication.java
@@ -1,7 +1,6 @@
 package com.softpro.dnaig.softpro;
 
 import javafx.application.Application;
-import javafx.fxml.FXMLLoader;
 import javafx.scene.Scene;
 import javafx.scene.control.Button;
 import javafx.scene.layout.VBox;
@@ -11,7 +10,7 @@ import java.io.IOException;
 
 public class MainApplication extends Application {
 
-    final int screenres = 1024;
+    final double screenres = 1024;
     @Override
     public void start(Stage stage) throws IOException {
         //FXMLLoader fxmlLoader = new FXMLLoader(MainApplication.class.getResource("hello-view.fxml"));
